Ignore file extension in package imports
Unfortunately the rule is not smart enough to distinguish relative file
imports in a node_modules package from import paths that are specified
in a node_modules package.json exports map.

In this case the rule reported `@effect/schema/Schema` that it requires
an extension. However this is incorrect because this path in specified
in the exports map and points to a different dist folder. While the path
would also exist on disk as it is, it is still incorrect since it is a
different (non-compiled) file. Adding the extension ends up in the
export path not being resolved, as it omits the extension there.