Compare tarball contents instead of gzipped shasum
Unfortunately zlib, which is used for the gzip compression of tarballs,
does not produce reproducible results. So depending on the exact version
of zlib or even on which operating system it runs, it might produce
different output. Therefore we will just give up on the approach to
compare the shasum, instead we download the tarball from the registry
and compare its contents. This has the advantage to we can also produce
diffs in the future.